refactor(dep-check): move core profiles into Microsoft preset
