+++ .changeset/moody-suits-sort.md
@@ -1,0 +1,5 @@
+---
+"@rnx-kit/dep-check": none
+---
+
+Move core profiles into Microsoft preset
