+++ packages/dep-check/src/presets/microsoft/profile-0.63.ts
@@ -1,116 +1,117 @@
-import type { Profile, Package } from "../types";
-import profile_0_65 from "./profile-0.65";
+import type { Package, Profile } from "../../types";
+import profile62 from "./profile-0.62";
 
 const reactNative: Package = {
   name: "react-native",
-  version: "^0.66.0-0",
+  version: "^0.63.2",
   capabilities: ["react"],
 };
 
 const profile: Profile = {
-  ...profile_0_65,
+  ...profile62,
+
   react: {
     name: "react",
-    version: "17.0.2",
+    version: "16.13.1",
   },
   "react-dom": {
     name: "react-dom",
-    version: "17.0.2",
+    version: "16.13.1",
     capabilities: ["react"],
   },
   "react-test-renderer": {
     name: "react-test-renderer",
-    version: "17.0.2",
+    version: "16.13.1",
     capabilities: ["react"],
     devOnly: true,
   },
+
   core: reactNative,
   "core-android": reactNative,
   "core-ios": reactNative,
   "core-macos": {
     name: "react-native-macos",
-    version: "^0.66.0-0",
+    version: "^0.63.0",
     capabilities: ["react"],
   },
   "core-windows": {
     name: "react-native-windows",
-    version: "^0.66.0-0",
+    version: "^0.63.0",
     capabilities: ["core"],
   },
 
-  animation: {
-    name: "react-native-reanimated",
-    version: "^2.2.3",
+  "babel-preset-react-native": {
+    name: "metro-react-native-babel-preset",
+    version: "^0.59.0",
+    devOnly: true,
+  },
+  "floating-action": {
+    name: "react-native-floating-action",
+    version: "^1.21.0",
+  },
+  gestures: {
+    name: "react-native-gesture-handler",
+    version: "^1.10.3",
   },
   hermes: {
     name: "hermes-engine",
-    version: "~0.9.0",
-  },
-  "babel-preset-react-native": {
-    name: "metro-react-native-babel-preset",
-    version: "^0.66.2",
-    devOnly: true,
-  },
-  html: {
-    name: "react-native-render-html",
-    version: "^6.1.0",
+    version: "~0.5.0",
   },
   jest: {
     name: "jest",
-    version: "^26.6.3",
+    version: "^24.9.0",
     devOnly: true,
-  },
-  "lazy-index": {
-    name: "@rnx-kit/react-native-lazy-index",
-    version: "^2.1.7",
   },
   metro: {
     name: "metro",
-    version: "^0.66.2",
+    version: "^0.59.0",
     devOnly: true,
   },
   "metro-config": {
     name: "metro-config",
-    version: "^0.66.2",
+    version: "^0.59.0",
     devOnly: true,
   },
   "metro-core": {
     name: "metro-core",
-    version: "^0.66.2",
+    version: "^0.59.0",
     devOnly: true,
   },
   "metro-react-native-babel-transformer": {
     name: "metro-react-native-babel-transformer",
-    version: "^0.66.2",
+    version: "^0.59.0",
     devOnly: true,
   },
   "metro-resolver": {
     name: "metro-resolver",
-    version: "^0.66.2",
+    version: "^0.59.0",
     devOnly: true,
   },
   "metro-runtime": {
     name: "metro-runtime",
-    version: "^0.66.2",
+    version: "^0.59.0",
     devOnly: true,
   },
   "navigation/native": {
     name: "@react-navigation/native",
-    version: "^6.0.8",
+    version: "^5.9.4",
   },
   "navigation/stack": {
     name: "@react-navigation/stack",
-    version: "^6.2.0",
+    version: "^5.14.4",
     capabilities: ["navigation/native"],
+  },
+  "safe-area": {
+    name: "react-native-safe-area-context",
+    version: "^3.2.0",
   },
   screens: {
     name: "react-native-screens",
-    version: "^3.9.0",
+    version: "^2.18.1",
   },
-  "test-app": {
-    name: "react-native-test-app",
-    version: "^1.0.6",
-    devOnly: true,
+  webview: {
+    name: "react-native-webview",
+    version: "^11.4.2",
   },
 };
