+++ packages/dep-check/src/presets/microsoft/profile-0.65.ts
@@ -1,116 +1,127 @@
-import type { Profile, Package } from "../types";
-import profile62 from "./profile-0.62";
+import type { Package, Profile } from "../../types";
+import profile_0_64 from "./profile-0.64";
 
 const reactNative: Package = {
   name: "react-native",
-  version: "^0.63.2",
+  version: "^0.65.0-0",
   capabilities: ["react"],
 };
 
 const profile: Profile = {
-  ...profile62,
-
+  ...profile_0_64,
   react: {
     name: "react",
-    version: "16.13.1",
+    version: "17.0.2",
   },
   "react-dom": {
     name: "react-dom",
-    version: "16.13.1",
+    version: "17.0.2",
     capabilities: ["react"],
   },
   "react-test-renderer": {
     name: "react-test-renderer",
-    version: "16.13.1",
+    version: "17.0.2",
     capabilities: ["react"],
     devOnly: true,
   },
+
   core: reactNative,
   "core-android": reactNative,
   "core-ios": reactNative,
   "core-macos": {
     name: "react-native-macos",
-    version: "^0.63.0",
+    version: "^0.65.0-0",
     capabilities: ["react"],
   },
   "core-windows": {
     name: "react-native-windows",
-    version: "^0.63.0",
+    version: "^0.65.0-0",
     capabilities: ["core"],
   },
 
-  "floating-action": {
-    name: "react-native-floating-action",
-    version: "^1.21.0",
-  },
-  gestures: {
-    name: "react-native-gesture-handler",
-    version: "^1.10.3",
-  },
-  hermes: {
-    name: "hermes-engine",
-    version: "~0.5.0",
+  animation: {
+    name: "react-native-reanimated",
+    version: "^2.2.1",
   },
   "babel-preset-react-native": {
     name: "metro-react-native-babel-preset",
-    version: "^0.59.0",
+    version: "^0.66.0",
     devOnly: true,
+  },
+  clipboard: {
+    name: "@react-native-clipboard/clipboard",
+    version: "^1.9.0",
+  },
+  "datetime-picker": {
+    name: "@react-native-community/datetimepicker",
+    version: "^3.5.2",
+  },
+  filesystem: {
+    name: "react-native-fs",
+    version: "^2.18.0",
+  },
+  "floating-action": {
+    name: "react-native-floating-action",
+    version: "^1.22.0",
+  },
+  hermes: {
+    name: "hermes-engine",
+    version: "~0.8.1",
+  },
+  jest: {
+    name: "jest",
+    version: "^26.6.3",
+    devOnly: true,
+  },
+  "masked-view": {
+    name: "@react-native-masked-view/masked-view",
+    version: "^0.2.6",
   },
   metro: {
     name: "metro",
-    version: "^0.59.0",
+    version: "^0.66.0",
     devOnly: true,
   },
   "metro-config": {
     name: "metro-config",
-    version: "^0.59.0",
+    version: "^0.66.0",
     devOnly: true,
   },
   "metro-core": {
     name: "metro-core",
-    version: "^0.59.0",
+    version: "^0.66.0",
     devOnly: true,
   },
   "metro-react-native-babel-transformer": {
     name: "metro-react-native-babel-transformer",
-    version: "^0.59.0",
+    version: "^0.66.0",
     devOnly: true,
   },
   "metro-resolver": {
     name: "metro-resolver",
-    version: "^0.59.0",
+    version: "^0.66.0",
     devOnly: true,
   },
   "metro-runtime": {
     name: "metro-runtime",
-    version: "^0.59.0",
+    version: "^0.66.0",
     devOnly: true,
   },
-  "navigation/native": {
-    name: "@react-navigation/native",
-    version: "^5.9.4",
+  modal: {
+    name: "react-native-modal",
+    version: "^13.0.0",
   },
-  "navigation/stack": {
-    name: "@react-navigation/stack",
-    version: "^5.14.4",
-    capabilities: ["navigation/native"],
-  },
-  "safe-area": {
-    name: "react-native-safe-area-context",
-    version: "^3.2.0",
+  netinfo: {
+    name: "@react-native-community/netinfo",
+    version: "^7.0.0",
   },
   screens: {
     name: "react-native-screens",
-    version: "^2.18.1",
+    version: "^3.7.0",
   },
   webview: {
     name: "react-native-webview",
-    version: "^11.4.2",
-  },
-  jest: {
-    name: "jest",
-    version: "^24.9.0",
-    devOnly: true,
+    version: "^11.13.0",
   },
 };
