+++ packages/dep-check/src/profiles.ts
@@ -6,35 +6,17 @@ import semverValid from "semver/functions/valid";
 import semverIntersects from "semver/ranges/intersects";
 import semverValidRange from "semver/ranges/valid";
 import { keysOf } from "./helpers";
-import profile_0_61 from "./profiles/profile-0.61";
-import profile_0_62 from "./profiles/profile-0.62";
-import profile_0_63 from "./profiles/profile-0.63";
-import profile_0_64 from "./profiles/profile-0.64";
-import profile_0_65 from "./profiles/profile-0.65";
-import profile_0_66 from "./profiles/profile-0.66";
-import profile_0_67 from "./profiles/profile-0.67";
-import profile_0_68 from "./profiles/profile-0.68";
+import { default as defaultPreset } from "./presets/microsoft";
 import type {
   MetaPackage,
   Package,
   Profile,
+  ProfileMap,
   ProfilesInfo,
   ProfileVersion,
 } from "./types";
 
 type Capabilities = Record<string, MetaPackage | Package>;
-type ProfileMap = Record<ProfileVersion, Profile>;
-
-export const defaultProfiles: Readonly<ProfileMap> = {
-  "0.61": profile_0_61,
-  "0.62": profile_0_62,
-  "0.63": profile_0_63,
-  "0.64": profile_0_64,
-  "0.65": profile_0_65,
-  "0.66": profile_0_66,
-  "0.67": profile_0_67,
-  "0.68": profile_0_68,
-};
 
 function getVersionComparator(
   versionOrRange: string
@@ -65,7 +47,7 @@ function isValidProfileMap(
 }
 
 function isValidProfileVersion(v: string): v is ProfileVersion {
-  return v in defaultProfiles;
+  return v in defaultPreset;
 }
 
 export function loadCustomProfiles(
@@ -108,7 +90,7 @@ export function loadCustomProfiles(
       false
     );
     if (hasCommonCapabilities) {
-      const allVersions = Object.keys(defaultProfiles) as ProfileVersion[];
+      const allVersions = Object.keys(defaultPreset) as ProfileVersion[];
       return allVersions.reduce<
         Record<string, Record<string, MetaPackage | Package>>
       >((expandedProfiles, version) => {
@@ -140,7 +122,7 @@ export function getProfileVersionsFor(
   }
 
   const isSatifisedBy = getVersionComparator(reactVersionRange);
-  const allVersions = keysOf(defaultProfiles);
+  const allVersions = keysOf(defaultPreset);
   return allVersions.reduce<ProfileVersion[]>((profiles, version) => {
     if (isSatifisedBy(version)) {
       profiles.push(version);
@@ -155,7 +137,7 @@ export function getProfilesFor(
 ): Profile[] {
   const customProfiles = loadCustomProfiles(customProfilesPath);
   const profiles = getProfileVersionsFor(reactVersionRange).map((version) => ({
-    ...defaultProfiles[version],
+    ...defaultPreset[version],
     ...customProfiles[version],
   }));
   if (profiles.length === 0) {
